updated to 3.0
