+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
 }
 
 group 'me.bristermitten'
-version '2.0.2'
+version '3.0.1'
 
 sourceCompatibility = targetCompatibility = JavaVersion.VERSION_1_8
