+++ src/main/kotlin/dev/triumphteam/gui/components/EnchantmentContainer.kt
@@ -1,14 +1,15 @@
-package me.mattstudios.mfgui.gui.components
+package dev.triumphteam.gui.components
 
+import dev.triumphteam.gui.builder.item.ItemBuilder
 import org.bukkit.enchantments.Enchantment
 
 class EnchantmentContainer(private val itemBuilder: ItemBuilder) {
     infix fun Enchantment.level(level: Int) {
-        itemBuilder.addEnchantment(this, level)
+        itemBuilder.enchant(this, level)
     }
 
     fun add(enchantment: Enchantment) {
-        itemBuilder.addEnchantment(enchantment)
+        itemBuilder.enchant(enchantment)
     }
 }
