+++ src/main/kotlin/dev/triumphteam/gui/guis/GuiItems.kt
@@ -1,6 +1,5 @@
-package me.mattstudios.mfgui.gui.guis
+package dev.triumphteam.gui.guis
 
-import dev.triumphteam.gui.guis.GuiItem
 import org.bukkit.event.inventory.InventoryClickEvent
 import org.bukkit.inventory.ItemStack
