+++ src/main/kotlin/dev/triumphteam/gui/guis/ItemsContainer.kt
@@ -1,7 +1,4 @@
-package me.mattstudios.mfgui.gui.guis
-
-import dev.triumphteam.gui.guis.BaseGui
-import dev.triumphteam.gui.guis.GuiItem
+package dev.triumphteam.gui.guis
 
 class ItemsContainer(private val gui: BaseGui) {
     operator fun get(slot: Int): GuiItem? {
